Add Icon in Current Weather
